requirements.txt
asgiref==3.3.4
Django==3.2.3
django-ninja==0.13.0
gunicorn==20.1.0
pydantic==1.8.2
pytz==2021.1
pyzbar==0.1.8
segno==1.3.3
sqlparse==0.4.1
typing-extensions==3.10.0.0
